Fixed the layering of the GetProperties call for both Drilling and Adaptive classes.  They were calling the HeeksObj::GetProperties() method directly where they should have been calling COp::GetProperties and letting COp::GetProperties call the HeeksObj::GetProperties method
Added execution_order parameter to the COp class although it doesn't do us much good yet.  I
think we need to rethink the definition of the Program::m_operations variable.  I think it
should be a listof COp objects rather than a list of HeeksObj objects.  I've left the
execution order parameters in the code because; a) it's easy to remove and b) it's no worse
than the 'comments' parameter that is also shared between all COp objects.

Fixed problem with Drilling object where, if the object title is updated, the Operations
list names were not being updated immediately.