Added execution order to COp class so that all NC operations inherit it.
Added tool_number to COp class so that all NC operations inherit it.

Changed NC generation loops to only issue a 'select tool' command if the tool number is greater than zero and
if the tool number is different from what it was in the previous run through the loop (previous operation).

I also sort the operations by execution order number before running through the various operations.

From a previous change, the comment values associated with all operations (COp) now appear in the properties
window as well as appearing in the generated GCode as a comment.