Added CorrelationTool class although I have commented it out for now.  It's not ready and there are some basic problems with it.  The first is that one of the uses I had intended for it was to recognise similar Sketch objects (holes within pads generated from edge detection in InkScape).  The idea was to select one of the small holes drawn at the centre of the pads and have the CorrelationTool find other graphical elements that were similar enough to that selected as to remember them.  I could then use this CorrelationTool object as an object itself upon which to place a drilling object.
It sort of works but only in a transient form.  Since a Sketch object is based on ObjList, it
doesn't include an m_id value.  This means that the ID numbers assigned to Sketch objects are
not maintained between file close/open sessions.  This means that the reference elements I'm
using (kept by type and id) are gone when I re-open the data file.

If I do it all within the same session it sort of works.  With a handful of objects it finds
them (although it's not very discerning yet).  With a reasonably small PCB picture, it finds 488
similar objects before faulting with a segmentation violation.  (I know what you're thinking.
They all do that when they're new.  They will loosen up with wear).

For the last two days I have checked in changes that Dan has had to cleanup.  I have commented
out the 'kickoff' points in HeeksCNC.cpp for this new class until I can get it working.  I didn't
want to loose the work though.

I should also mention that I'm not sure this class will eventually fit into the model used
by the rest of the CAD package.  It's really an intelegent 'selection' tool but it's placed
into the CNC tools list.  I'm not sure it makes sense.  In any case, we can discard it later
when a better option is available (assuming there isn't one now)