Changed CounterBore to be based on CDepthOp rather than COp.
Changed CDepthOpParams::set_initial_values() method so that the tool number (and its diameter) is
carried through if it's present.

I haven't removed the tool_number and tool_diameter values from CDepthOpParams yet but I think they should be.  It just seemed
a little presumptuous of me to make this change.  Any suggestions as to whether this is reasonable?