Added an option for the user to select a cutting tool object as well when creating an Adaptive Clearing object.  If it's found, the toolflatrad and toolcornerrad values will be taken from the cutting tool object instead of the values found in the Adaptive Clearing object's parameters.
Again, if a cutting tool is referenced during the Adaptive Clearing object's creation,
the values of toolfatrad and toolcornerrad will be set based on the cutting tool's
values.  This is just in case the user deletes the cutting tool object before generating
the GCode.  I think these two parameters should be removed from the Adaptive class
but I'm not ready to impose this wish on every one else's data files.