Linked the kurve code into the HeeksCNC project so that we could determine the roll_on coordinate for each sketch in a profile.  At the moment we're not doing anything with these coordinates but there are enough changes that I want to check them in.  I don't believe I have broken anything. Right now the application should work the same way as it always did.  The extra calls into the kurve routines are being made in parallel with existing code.
I have modified the kurve code so that it presents a Python wrapper when compiled with its
own make file (and KURVE_PYTHON_INTERFACE macro defined) and a C++ wrapper when compiled
with the HeeksCNC makefile.  I haven't converted all the methods/routines.  I have only done
as many as are needed to calculate the roll-on points.

The next step is to add an optional drilling operation to the profile operation.