Changed ZigZag GCode generation routine to use the child objects as references rather than the list of solids (by ID).
Added a 'Reset Boundary' menu option to ZigZag operations so that, if the reference solids are changed manually, the area used for the ZigZag operation could be automatically reset.  This is not done automatically as this boundary is defined as a parameter.  If it had been reset manually then our code should not overwrite these settings.