find data files in the DLL dir, not the CWD
Before this commit, the plugin used to look for its default data files
in ".", the current working directory.  It was expecting the HeeksCAD
application to chdir to the plugin's install directory before invoking
the OnStartUp method.

After this commit, the plugin determines its install dir and look in
there, instead of making assumptions about where its working directory is.