Fixed a problem where the OutputCanvas::m_textCtrl (wxTextCtrl) could only hold 30k byte strings as a maximum.  Although there is a wxTextCtrl::SetMaxLength() call, it won't accept values larger than the underlying window objects support.  On win32, this is 62573 bytes.  This became a problem when I tried to use the contour operation to machine a PCB.  The combination of the verbose python script generated by the contour (and inlay) operations combined with a large number of small sketches (traces) made the wxTextCtrl limitation a problem.
The code now aggregates the python program in a 'Python' class that is based on wxString.  It overloads the operator<<(const double value) method so that the comma/dot problem we had recently does not re-surface.  Only at the very end does it attempt to put the python program into the wxTextCtrl object.  If there is not enough space, it adds a warning message indicating such along with the path where the user can access the python program directly.  The actual python program is still written to the external file and executed so the gcode still comes out.

I know this is a very large change but I like the fact that the python script generation methods return the python program rather than just adding it to a global variable.  I have used this to nest pocket operations within Inlay operations.  Having said this, I am happy to be overridden if others believe this is not the best solution.  We can rollback this change.