Reinstated CNCPoint.cpp and CNCPoint.h into the heekscnc/src directory.
Removed the 'find' command in the Makefile that was added to help find the way back from the interfaces directory to the heekscnc/src directory.

Fixed bugs in Inlay operation where the mountain walls were being formed from the bottom up.

Removed toolpath in Inlay operation that was along the top surface for both the Female and Male halves.