Fixed problem in edge probing operation.
Added option to have two private fixtures added to the Inlay operation so that the male and female halves could be machined in separate fixtures.

Added 'touch-off point' and 'touch-off description' fields to the Fixture objects.  This allows an X,Y rapid movement when moving between two fixtures.  It also allows a comment to be added to the HeeksCNC model indicating how the fixture should be setup.  eg: 'set 0,0 point to bottom left corner'.  This is added as a comment in the resultant GCode.