Added option into the spindle speed setting routine to recognise that this is a tapping tool and to reduce the spindle speed to 500 rpm accordingly.
Changed the code around so that a tapping tool's spindle
speed could be set without pre-selecting a material name
or hardness.

Added post-processing step to when the operator manually
changes the spindle speed value.  It now re-calls the SetFeeds() routine
to recalculate the feedrate based on the new spindle speed.

Changed the FeedRate() routine to handle a tapping tool's
pitch and use both it and the spindle speed to set the feed
rate.